Matrix: fix Jiti runtime API boundary
(cherry picked from commit ff6541f69d2e6cd88424953b13a43a20fa7aefb9)
