+++ extensions/matrix/runtime-api.ts
@@ -1,74 +1,4 @@
-export {
-  GROUP_POLICY_BLOCKED_LABEL,
-  MarkdownConfigSchema,
-  PAIRING_APPROVED_MESSAGE,
-  ToolPolicySchema,
-  buildChannelConfigSchema,
-  buildChannelKeyCandidates,
-  buildProbeChannelStatusSummary,
-  buildSecretInputSchema,
-  collectStatusIssuesFromLastError,
-  compileAllowlist,
-  createActionGate,
-  createReplyPrefixOptions,
-  createScopedPairingAccess,
-  createTypingCallbacks,
-  dispatchReplyFromConfigWithSettledDispatcher,
-  evaluateGroupRouteAccessForPolicy,
-  fetchWithSsrFGuard,
-  formatAllowlistMatchMeta,
-  formatLocationText,
-  hasConfiguredSecretInput,
-  issuePairingChallenge,
-  jsonResult,
-  logInboundDrop,
-  logTypingFailure,
-  mergeAllowlist,
-  normalizeResolvedSecretInputString,
-  normalizeSecretInputString,
-  normalizeStringEntries,
-  readNumberParam,
-  readReactionParams,
-  readStoreAllowFromForDmPolicy,
-  readStringParam,
-  resolveAllowlistProviderRuntimeGroupPolicy,
-  resolveChannelEntryMatch,
-  resolveCompiledAllowlistMatch,
-  resolveControlCommandGate,
-  resolveDefaultGroupPolicy,
-  resolveDmGroupAccessWithLists,
-  resolveInboundSessionEnvelopeContext,
-  resolveRuntimeEnv,
-  resolveSenderScopedGroupPolicy,
-  runPluginCommandWithTimeout,
-  summarizeMapping,
-  toLocationContext,
-  warnMissingProviderGroupPolicyFallbackOnce,
-  DEFAULT_ACCOUNT_ID,
-} from "openclaw/plugin-sdk/matrix";
-export { createAccountListHelpers } from "openclaw/plugin-sdk/account-helpers";
-export type {
-  AllowlistMatch,
-  BaseProbeResult,
-  ChannelDirectoryEntry,
-  ChannelGroupContext,
-  ChannelMessageActionAdapter,
-  ChannelMessageActionContext,
-  ChannelMessageActionName,
-  ChannelOutboundAdapter,
-  ChannelPlugin,
-  ChannelResolveKind,
-  ChannelResolveResult,
-  ChannelToolSend,
-  DmPolicy,
-  GroupPolicy,
-  GroupToolPolicyConfig,
-  MarkdownTableMode,
-  NormalizedLocation,
-  PluginRuntime,
-  PollInput,
-  ReplyPayload,
-  RuntimeEnv,
-  RuntimeLogger,
-  SecretInput,
-} from "openclaw/plugin-sdk/matrix";
+// Keep the external runtime API light so Jiti callers can resolve Matrix config
+// helpers without traversing the full plugin-sdk/runtime graph.
+export * from "./src/auth-precedence.js";
+export * from "./helper-api.js";
