+++ extensions/matrix/src/runtime-api.ts
@@ -1,13 +1,2 @@
 export * from "openclaw/plugin-sdk/matrix";
-export {
-  assertHttpUrlTargetsPrivateNetwork,
-  closeDispatcher,
-  createPinnedDispatcher,
-  resolvePinnedHostnameWithPolicy,
-  ssrfPolicyFromAllowPrivateNetwork,
-  type LookupFn,
-  type SsrFPolicy,
-} from "openclaw/plugin-sdk/infra-runtime";
-// Keep auth-precedence available internally without re-exporting helper-api
-// twice through both plugin-sdk/matrix and ../runtime-api.js.
-export * from "./auth-precedence.js";
+export * from "../runtime-api.js";
