+++ extensions/matrix/thread-bindings-runtime.ts
@@ -1,0 +1,4 @@
+export {
+  setMatrixThreadBindingIdleTimeoutBySessionKey,
+  setMatrixThreadBindingMaxAgeBySessionKey,
+} from "./src/matrix/thread-bindings-shared.js";
